YL00-1000#Update the ui

+++ blog/src/app/admin/admin.service.ts
@@ -1,9 +1,4 @@
-import {
-  Page,
-  TableMetadata,
-  ColumnMetadata,
-  SortColumn,
-} from './../model/metadata.model';
+import { Page, TableMetadata, SortColumn } from './../model/metadata.model';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Blog } from '../blog';
@@ -22,12 +17,12 @@ export class AdminService extends Blog {
   //Get all the table names
   getBlog(metadata: string) {
     let apiEndPoint = this.apiEndPoint + metadata;
-    // if (this.checkJwtToken()) {
-    //   this.router.navigate(['/login']);
-    // } else {
-    this.addTokenInHeader();
-    return this.http.get(apiEndPoint, { headers: this.headers });
-    // }
+    if (this.checkJwtToken()) {
+      this.router.navigate(['/login']);
+    } else {
+      this.addTokenInHeader();
+      return this.http.get(apiEndPoint, { headers: this.headers });
+    }
   }
 
   //get all column details for the tables
@@ -46,19 +41,19 @@ export class AdminService extends Blog {
     metadata: string,
     page?: Page,
     filter?: string,
-    columns?: ColumnMetadata[]
+    columns?: any
   ) {
     let apiEndPoint = this.apiEndPoint + '/' + metadata;
     let params = this.getSearchParamString(page, filter, columns);
     if (params) {
       apiEndPoint = apiEndPoint + params;
     }
-    // if(this.checkJwtToken()){
-    //   this.router.navigate(['/login']);
-    // }else{
-    this.addTokenInHeader();
-    return this.http.get(apiEndPoint, { headers: this.headers });
-    // }
+    if (this.checkJwtToken()) {
+      this.router.navigate(['/login']);
+    } else {
+      this.addTokenInHeader();
+      return this.http.get(apiEndPoint, { headers: this.headers });
+    }
   }
 
   //Get all the data for the table in pagignated way and sorted
@@ -68,55 +63,55 @@ export class AdminService extends Blog {
     if (params) {
       apiEndPoint = apiEndPoint + params;
     }
-    // if(this.checkJwtToken()){
-    //   this.router.navigate(['/login']);
-    // }else{
-    this.addTokenInHeader();
-    return this.http.get(apiEndPoint, { headers: this.headers });
-    // }
+    if (this.checkJwtToken()) {
+      this.router.navigate(['/login']);
+    } else {
+      this.addTokenInHeader();
+      return this.http.get(apiEndPoint, { headers: this.headers });
+    }
   }
 
   //fetch one row based on id
   getData(metadata: string, data: any) {
     let apiEndPoint = this.apiEndPoint + '/' + metadata + '/' + data.id;
-    // if(this.checkJwtToken()){
-    //   this.router.navigate(['/login']);
-    // }else{
-    this.addTokenInHeader();
-    return this.http.get(apiEndPoint, { headers: this.headers });
-    // }
+    if (this.checkJwtToken()) {
+      this.router.navigate(['/login']);
+    } else {
+      this.addTokenInHeader();
+      return this.http.get(apiEndPoint, { headers: this.headers });
+    }
   }
 
   //add data
   addData(metadata: string, data: any) {
     let apiEndPoint = this.apiEndPoint + '/' + metadata;
-    // if(this.checkJwtToken()){
-    //   this.router.navigate(['/login']);
-    // }else{
-    this.addTokenInHeader();
-    return this.http.post(apiEndPoint, data, { headers: this.headers });
-    // }
+    if (this.checkJwtToken()) {
+      this.router.navigate(['/login']);
+    } else {
+      this.addTokenInHeader();
+      return this.http.post(apiEndPoint, data, { headers: this.headers });
+    }
   }
 
   //update data
   updateData(metadata: string, data: any) {
     let apiEndPoint = this.apiEndPoint + '/' + metadata + '/' + data.id;
-    // if(this.checkJwtToken()){
-    //   this.router.navigate(['/login']);
-    // }else{
-    this.addTokenInHeader();
-    return this.http.put(apiEndPoint, data, { headers: this.headers });
-    // }
+    if (this.checkJwtToken()) {
+      this.router.navigate(['/login']);
+    } else {
+      this.addTokenInHeader();
+      return this.http.put(apiEndPoint, data, { headers: this.headers });
+    }
   }
 
   //delete data
   deleteData(metadata: string, data: any) {
     let apiEndPoint = this.apiEndPoint + '/' + metadata + '/' + data.id;
-    // if(this.checkJwtToken()){
-    //   this.router.navigate(['/login']);
-    // }else{
-    this.addTokenInHeader();
-    return this.http.delete(apiEndPoint, { headers: this.headers });
-    // }
+    if (this.checkJwtToken()) {
+      this.router.navigate(['/login']);
+    } else {
+      this.addTokenInHeader();
+      return this.http.delete(apiEndPoint, { headers: this.headers });
+    }
   }
 }
